refactor: convert attackDialog to AppV2
