+++ eslint.config.mjs
@@ -50,6 +50,8 @@ export default [
     rules: {
       "@typescript-eslint/no-explicit-any": "off",
       "@typescript-eslint/no-empty-object-type": "off",
+      "@typescript-eslint/no-unused-vars": "warn",
+      // "no-unused-vars": "warn", TODO enable when dataModels are used. Now it overflows with warnings.
 
       "prettier/prettier": "error",
       "no-unused-expressions": [2, { allowTernary: true }],
