+++ src/actors/sheet-parts/combat-tab/combat.hbs
@@ -26,7 +26,7 @@
 {{/if}}
 
 <div class="flex-row flex-align-center">
-  <label data-damage-roll="{{system.attributes.damageBonus}}" data-damage-roll-heading="RQG.Actor.Combat.DamageBonus">>{{localize "RQG.Actor.Combat.DamageBonus"}} {{system.attributes.damageBonus}}</label>
+  <label data-damage-roll="{{system.attributes.damageBonus}}" data-damage-roll-heading="RQG.Actor.Combat.DamageBonus">{{localize "RQG.Actor.Combat.DamageBonus"}} {{system.attributes.damageBonus}}</label>
 </div>
 <div class="grid combat item-list">
   <div class="headings"></div>
