+++ src/i18n/en/uiContent.json
@@ -19,6 +19,9 @@
     },
     "Actor": {
       "character": "Character"
+    },
+    "ChatMessage": {
+      "combat": "Combat"
     }
   },
   "COMBAT": {
